+++ src/block-kit/Blocks.tsx
@@ -32,7 +32,10 @@ const basicBlocks = ['actions', 'context', 'divider', 'image', 'section']
 
 // Blocks
 knownBlocks.set(undefined, [...basicBlocks, 'file'])
-knownActions.set(undefined, actionTypes.filter(t => t !== 'radio_buttons'))
+knownActions.set(
+  undefined,
+  actionTypes.filter(t => t !== 'radio_buttons')
+)
 knownSectionAccessories.set(
   undefined,
   sectionAccessoryTypes.filter(t => t !== 'radio_buttons')
@@ -45,14 +48,8 @@ knownBlocks.set(InternalBlockType.Modal, [
   internalHiddenType,
   internalSubmitType,
 ])
-knownActions.set(
-  InternalBlockType.Modal,
-  actionTypes.filter(t => t !== 'radio_buttons')
-)
-knownSectionAccessories.set(
-  InternalBlockType.Modal,
-  sectionAccessoryTypes.filter(t => t !== 'radio_buttons')
-)
+knownActions.set(InternalBlockType.Modal, [...actionTypes])
+knownSectionAccessories.set(InternalBlockType.Modal, [...sectionAccessoryTypes])
 
 // Home
 knownBlocks.set(InternalBlockType.Home, basicBlocks)
@@ -128,6 +125,6 @@ export const Blocks: JSXSlack.FC<BlocksProps> = props => {
   return <ArrayOutput>{normalized}</ArrayOutput>
 }
 
-export const BlocksInternal: JSXSlack.FC<
-  BlocksProps & { [blockTypeSymbol]?: InternalBlockType }
-> = props => Blocks(props)
+export const BlocksInternal: JSXSlack.FC<BlocksProps & {
+  [blockTypeSymbol]?: InternalBlockType
+}> = props => Blocks(props)